corrections after ux review
