+++ app/screens/edit_profile/components/email_field.tsx
@@ -3,20 +3,13 @@
 
 import React, {RefObject} from 'react';
 import {useIntl} from 'react-intl';
+import {Text, View} from 'react-native';
 
 import {FloatingTextInputRef} from '@components/floating_text_input_label';
+import {changeOpacity, makeStyleSheetFromTheme} from '@utils/theme';
+import {typography} from '@utils/typography';
 
 import Field from './field';
-
-type EmailSettingsProps = {
-    authService: string;
-    email: string;
-    fieldRef: RefObject<FloatingTextInputRef>;
-    onChange: (fieldKey: string, value: string) => void;
-    onFocusNextField: (fieldKey: string) => void;
-    isDisabled: boolean;
-    label: string;
-}
 
 const services: Record<string, string> = {
     gitlab: 'GitLab',
@@ -26,6 +19,30 @@ const services: Record<string, string> = {
     saml: 'SAML',
 };
 
+const getStyleSheet = makeStyleSheetFromTheme((theme) => {
+    return {
+        container: {
+            marginTop: 8,
+        },
+        text: {
+            ...typography('Body', 75),
+            color: changeOpacity(theme.centerChannelColor, 0.5),
+        },
+    };
+});
+
+type EmailSettingsProps = {
+    authService: string;
+    email: string;
+    fieldRef: RefObject<FloatingTextInputRef>;
+    onChange: (fieldKey: string, value: string) => void;
+    onFocusNextField: (fieldKey: string) => void;
+    isDisabled: boolean;
+    label: string;
+    theme: Theme;
+    isTablet: boolean;
+}
+
 const EmailField = ({
     authService,
     email,
@@ -34,9 +51,12 @@ const EmailField = ({
     onFocusNextField,
     isDisabled,
     label,
+    theme,
+    isTablet,
 }: EmailSettingsProps) => {
     const intl = useIntl();
     const service = services[authService];
+    const style = getStyleSheet(theme);
 
     let fieldDescription: string;
 
@@ -50,22 +70,32 @@ const EmailField = ({
             defaultMessage: 'Email must be updated using a web client or desktop application.'}, {email, service});
     }
 
+    const descContainer = [style.container, {paddingHorizontal: isTablet ? 42 : 20}];
+
     return (
-        <Field
-            blurOnSubmit={false}
-            enablesReturnKeyAutomatically={true}
-            fieldDescription={fieldDescription}
-            fieldKey='email'
-            fieldRef={fieldRef}
-            isDisabled={isDisabled}
-            keyboardType='email-address'
-            label={label}
-            onFocusNextField={onFocusNextField}
-            onTextChange={onChange}
-            returnKeyType='next'
-            testID='edit_profile.text_setting.email'
-            value={email}
-        />
+        <>
+            <Field
+                blurOnSubmit={false}
+                enablesReturnKeyAutomatically={true}
+                fieldKey='email'
+                fieldRef={fieldRef}
+                isDisabled={isDisabled}
+                keyboardType='email-address'
+                label={label}
+                onFocusNextField={onFocusNextField}
+                onTextChange={onChange}
+                returnKeyType='next'
+                testID='edit_profile.text_setting.email'
+                value={email}
+            />
+            <View
+                style={descContainer}
+            >
+                <Text style={style.text}>{fieldDescription}</Text>
+            </View>
+
+        </>
+
     );
 };
